Dialog de Confirmação

+++ src/app/courses/containers/courses/courses.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfimationDialogComponent } from 'src/app/shared/components/confimation-dialog/confimation-dialog.component';
 import { ErroDialogComponent } from 'src/app/shared/components/erro-dialog/erro-dialog.component';
 
 import { Course } from '../../model/course';
@@ -54,16 +55,24 @@ export class CoursesComponent {
   }
 
   onRemove(course:Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open("Curso removido com sucesso!",'X',{
-          duration:5000,
-          verticalPosition: "top",
-          horizontalPosition: "center"
-        });
-      },
-      () => this.onError("Erro ao tentar remover curso!")
-    );
+    const dialogRef = this.dialog.open(ConfimationDialogComponent, {
+      data: "Tem certeza que deseja remover esse curso?",
+    });
+
+    dialogRef.afterClosed().subscribe((result:boolean )=> {
+      if(result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open("Curso removido com sucesso!",'X',{
+              duration:5000,
+              verticalPosition: "top",
+              horizontalPosition: "center"
+            });
+          },
+          () => this.onError("Erro ao tentar remover curso!")
+        );
+      }
+    });
   }
 }
